+++ base/application/service.py
@@ -9,6 +9,8 @@ import base.application.components
 from base.application.components import BaseHandler
 from base.application.components import DefaultRouteHandler
 from base.application.helpers.exceptions import MissingApplicationPort
+from base.application.helpers.exceptions import PreApplicationProcessConfigurationError
+from base.application.helpers.exceptions import PostApplicationProcessConfigurationError
 from base.application.helpers.importer import load_application
 from base.application.helpers.importer import load_orm
 
@@ -62,14 +64,7 @@ def _add_prefix(entries, prefix, svc_port):
     return _new_entries
 
 
-def engage(starter_path):
-
-    args = check_arguments()
-
-    entries = [(BaseHandler.__URI__, BaseHandler), ]
-    load_application(entries, args.port)
-    entries.append((r"/static/(.*)", tornado.web.StaticFileHandler,
-                    {"path": "{}/static".format(os.path.dirname(starter_path))}))
+def _engage(args, entries):
 
     svc_port = _get_svc_port()
     if not svc_port:
@@ -97,7 +92,84 @@ def engage(starter_path):
     tornado.ioloop.IOLoop.instance().start()
 
 
-if __name__ == '__main__':
+def _start_app_processes(app):
 
-    engage()
+    from base.common.utils import log
+    import subprocess
+
+    def _log(msg):
+        print(msg)
+        log.info(msg)
+    _log('Starting {}'.format(app[0]))
+    try:
+        if app[2]:  # if redirect output
+            p = subprocess.Popen(app[1], stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+        else:
+            p = subprocess.Popen(app[1])
+    except FileNotFoundError:
+        _msg = 'Can not find process {}'.format(app[0])
+        log.error(_msg)
+        raise
+    _log('{} started: PID {}'.format(app[0], p.pid))
+
+
+def _start_pre_app_processes():
+
+    # [(app_name, app_cmd_for_subprocess, redirect_output)]
+    for app in base.config.application_config.pre_app_processes:
+        _start_app_processes(app)
+
+    return True
+
+
+def _start_post_app_processes():
+
+    # [(app_name, app_cmd_for_subprocess, redirect_output)]
+    for app in base.config.application_config.post_app_processes:
+        _start_app_processes(app)
+
+    return True
+
+
+def _engage_with_process(args, entries):
+
+    from base.common.utils import log
+    if base.config.application_config.pre_app_processes is not None:
+
+        if type(base.config.application_config.pre_app_processes) != list:
+            log.error('Invalid type in configuration for pre application processes')
+            raise PreApplicationProcessConfigurationError('Pre application processes has to be a list')
+        if any([type(x) != tuple for x in base.config.application_config.pre_app_processes]) or \
+                any([len(x) != 3 for x in base.config.application_config.pre_app_processes]):
+            log.error('Invalid configuration for pre application processes. Processes has to be declared as a tuple of a name and a command')
+            raise PreApplicationProcessConfigurationError('Pre application processes has to be tuples in a list')
+
+        if not _start_pre_app_processes():
+            log.error('Could not start pre application processes')
+
+    if base.config.application_config.post_app_processes is not None:
+
+        if type(base.config.application_config.post_app_processes) != list:
+            log.error('Invalid type in configuration for post application processes')
+            raise PostApplicationProcessConfigurationError('Post application processes has to be a list')
+        if any([type(x) != tuple for x in base.config.application_config.post_app_processes]) or \
+                any([len(x) != 3 for x in base.config.application_config.post_app_processes]):
+            log.error('Invalid configuration for post application processes. Processes has to be declared as a tuple of a name and a command')
+            raise PostApplicationProcessConfigurationError('Post application processes has to be tuples in a list')
+
+        tornado.ioloop.IOLoop.current().add_callback(_start_post_app_processes)
+
+    _engage(args, entries)
+
+
+def engage(starter_path):
+
+    args = check_arguments()
+
+    entries = [(BaseHandler.__URI__, BaseHandler), ]
+    load_application(entries, args.port)
+    entries.append((r"/static/(.*)", tornado.web.StaticFileHandler,
+                    {"path": "{}/static".format(os.path.dirname(starter_path))}))
+
+    _engage_with_process(args, entries)