+++ base/config/application_config.py
@@ -46,3 +46,5 @@ static_uri = ''
 log_directory = './log'
 register_allowed_roles = None
 registrators_allowed_roles = None
+pre_app_processes = None    # [(app_name, app_cmd_for_subprocess, redirect_output)]
+post_app_processes = None   # [(app_name, app_cmd_for_subprocess, redirect_output)]